refactor(data-requests): replace dr pool vector with cw map w/ count

+++ packages/proxy/src/msg.rs
@@ -50,7 +50,7 @@ pub enum ProxyQueryMsg {
     #[returns(GetDataRequestsFromPoolResponse)]
     GetDataRequestsFromPool {
         position: Option<u128>,
-        limit: Option<u32>,
+        limit: Option<u128>,
     },
     #[returns(GetCommittedDataResultResponse)]
     GetCommittedDataResult { dr_id: Hash, executor: Addr },
